+++ src/schema.rs
@@ -7,8 +7,8 @@
 
 mod ddl_reader;
 
-pub(crate) use ddl_reader::rowid_alias_column;
 pub use ddl_reader::{
-    column_defs, column_type, read_schema, read_table_and_view_names, read_views, DdlError,
-    IndexSchema, IndexedColumn, TableSchema, ViewSchema,
+    column_defs, column_type, read_schema, read_schema_and_views, read_table_and_view_names,
+    read_views, rowid_alias_from_sql, DdlError, IndexSchema, IndexedColumn, TableSchema,
+    ViewSchema,
 };